pyproject.toml
[tool.poetry]
name = "docker"
version = "0.1.0"
description = ""
authors = ["Elizabeth <gresdiversaidade@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
mysqlclient = "^2.0.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
